Begin working with images
Started by adding a TableView for listing images ‘in’ the app. A button
and title are there, but need work. As of now the table view will list
any images prefixed with ‘ss_’ but nothing more. See ViewController for
how to reference objects that aren’t necessarily the entire screen.