Another feeble attempt to reduce memory usage
Passing only the name of the new image rather than the whole thing. The
image itself is still being passed to save it to the database which is
retained after the save, but that amounts to a single extra image being
held in memory. Looks to be about as far down as I can go.